+++ server/auth/types/authentication_type.ts
@@ -28,7 +28,7 @@ import {
 } from 'opensearch-dashboards/server';
 import { SecurityPluginConfigType } from '../..';
 import { SecuritySessionCookie } from '../../session/security_cookie';
-import { SecurityClient } from '../../backend/opendistro_security_client';
+import { SecurityClient } from '../../backend/opensearch_security_client';
 import {
   isMultitenantPath,
   resolveTenant,
