+++ server/auth/types/saml/routes.ts
@@ -21,7 +21,7 @@ import {
 } from '../../../../../../src/core/server';
 import { SecuritySessionCookie } from '../../../session/security_cookie';
 import { SecurityPluginConfigType } from '../../..';
-import { SecurityClient } from '../../../backend/opendistro_security_client';
+import { SecurityClient } from '../../../backend/opensearch_security_client';
 import { CoreSetup } from '../../../../../../src/core/server';
 import { validateNextUrl } from '../../../utils/next_url';
