+++ server/backend/opensearch_security_client.ts
