+++ server/backend/opensearch_security_configuration_plugin.ts
